Remove variable-size prototype, CRTFrame is now fixed-size
